+++ src/yt_game_records.c
@@ -317,7 +317,8 @@ yt_game_open(struct yt_game *game, enum yt_open_mode mode,
 		yt_game_close(game);
 		return false;
 	}
-	return yt_current_date_serial(&game->clock, game->config.epoch_year,
+	return yt_current_date_serial(&game->clock,
+	    (float)game->config.epoch_year,
 	    &game->today, &game->adjusted_year, error);
 }
